# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Questionnaire API",
    "version": "1.0.0",
    "description": "API for managing questionnaires and responses"
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "value": "Bearer <JWT token here>"
      }
    }
  },
  "paths": {
...
    "/mbti/questions": {
      "get": {
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Retrieve all MBTI-related questions.",
        "description": "Fetch a list of questions designed for MBTI assessment.",
        "tags": [
          "MBTI"
        ],
        "responses": {
          "200": {
            "description": "List of MBTI questions retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "question_id": {
                        "type": "integer",
                        "description": "ID of the question."
                      },
                      "question_text": {
                        "type": "string",
                        "description": "Text of the MBTI question."
                      },
                      "dimension": {
                        "type": "string",
                        "description": "Personality dimension (e.g., EI, SN, TF, JP)."
                      },
                      "direction": {
                        "type": "string",
                        "enum": [
                          "positive",
                          "negative"
                        ],
                        "description": "Direction of scoring for the dimension."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch MBTI questions."
          }
        }
      },
      "post": {
        "summary": "Create a new MBTI question.",
        "description": "Add a question for the Myers-Briggs Type Indicator (MBTI) assessment.",
        "tags": [
          "MBTI"
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "question_text",
                  "dimension",
                  "direction"
                ],
                "properties": {
                  "question_text": {
                    "type": "string",
                    "description": "The text of the MBTI question.",
                    "example": "You enjoy vibrant social events with lots of people."
                  },
                  "dimension": {
                    "type": "string",
                    "description": "The MBTI dimension the question is assessing.",
                    "enum": [
                      "EI",
                      "SN",
                      "TF",
                      "JP"
                    ],
                    "example": "EI"
                  },
                  "direction": {
                    "type": "string",
                    "description": "Indicates if the question has a positive or negative impact on the score.",
                    "enum": [
                      "positive",
                      "negative"
                    ],
                    "example": "positive"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MBTI question created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "question_id": {
                      "type": "integer",
                      "description": "The unique ID of the created question.",
                      "example": 1
                    },
                    "question_text": {
                      "type": "string",
                      "example": "You enjoy vibrant social events with lots of people."
                    },
                    "dimension": {
                      "type": "string",
                      "example": "EI"
                    },
                    "direction": {
                      "type": "string",
                      "example": "positive"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing fields."
          },
          "500": {
            "description": "Server error during question creation."
          }
        }
      }
    },
    "/mbti/submit": {
      "post": {
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Submit user's MBTI assessment answers.",
        "description": "Calculate MBTI type based on the user's answers and store it in the database.",
        "tags": [
          "MBTI"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "question_id": {
                          "type": "integer",
                          "description": "ID of the answered question."
                        },
                        "response": {
                          "type": "integer",
                          "description": "User's response to the question (1-5 scale)."
                        }
                      }
                    }
                  }
                },
                "required": [
                  "answers"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MBTI type calculated and saved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "User's calculated MBTI type (e.g., INTJ)."
                    },
                    "description": {
                      "type": "string",
                      "description": "Description of the calculated MBTI type."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing required fields."
          },
          "500": {
            "description": "Failed to submit MBTI answers."
          }
        }
      }
    },
    "/mbti/user-type": {
      "get": {
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get user's MBTI type.",
        "description": "Retrieve the MBTI type for the authenticated user.",
        "tags": [
          "MBTI"
        ],
        "responses": {
          "200": {
            "description": "User's MBTI type retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type_name": {
                      "type": "string",
                      "description": "MBTI type of the user (e.g., INTJ)."
                    },
                    "description": {
                      "type": "string",
                      "description": "Description of the user's MBTI type."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "MBTI type not found for user."
          },
          "500": {
            "description": "Failed to fetch user's MBTI type."
          }
        }
      }
    },
    "/mbti/admin-results": {
      "get": {
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get MBTI results for all members.",
        "description": "Retrieve MBTI types and descriptions for all members. Admin-only access.",
        "tags": [
          "MBTI"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination (default is 1)."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of records per page (default is 10)."
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Optional search term to filter by username or MBTI type."
          }
        ],
        "responses": {
          "200": {
            "description": "List of members and their MBTI results retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "user_id": {
                            "type": "integer",
                            "description": "User's ID."
                          },
                          "username": {
                            "type": "string",
                            "description": "User's name."
                          },
                          "type_name": {
                            "type": "string",
                            "description": "User's MBTI type."
                          },
                          "description": {
                            "type": "string",
                            "description": "Description of the MBTI type."
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin-only access."
          },
          "500": {
            "description": "Server error during retrieval."
          }
        }
      }
    },
    "/mbti/member-result": {
      "get": {
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get individual MBTI result.",
        "description": "Retrieve the authenticated user's MBTI type and description.",
        "tags": [
          "MBTI"
        ],
        "responses": {
          "200": {
            "description": "User's MBTI result retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type_name": {
                      "type": "string",
                      "description": "MBTI type of the user (e.g., INTJ)."
                    },
                    "description": {
                      "type": "string",
                      "description": "Description of the user's MBTI type."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "MBTI result not found for user."
          },
          "500": {
            "description": "Server error during retrieval."
          }
        }
      }
    },
...
  },
...
}